#5: Refactor
