--- src/data_classes/EnumDirections.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class Direction(Enum):
-    UP = 1
-    DOWN = 2
-    LEFT = 3
-    RIGHT = 4
