+++ src/data_classes/EnumEvent.py
@@ -1,12 +1,13 @@
 from enum import Enum
 
 
-class ControllerEvent(Enum):
+class Event(Enum):
     MOVE_UP = 1
     MOVE_DOWN = 2
     MOVE_LEFT = 3
     MOVE_RIGHT = 4
 
-    START = 5
-    RESTART = 6
-    EXIT = 7
+    START_MENU = 5
+    RUNNING = 6
+    GAME_OVER = 7
+
