--- src/data_classes/EnumPhases.py
@@ -1,7 +1,0 @@
-from enum import Enum
-
-
-class Phase(Enum):
-    MENU = 1
-    RUNNING = 2
-    GAME_OVER = 3
