+++ src/data_classes/GameState.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass, field
 from typing import List, Tuple
-from EnumPhases import Phase
+from data_classes.EnumEvent import Event
 
 MAP_SIZE = 30
 
@@ -16,4 +16,5 @@ class GameState:
     snake: List[Coord] = field(default_factory=lambda: [Coord(MAP_SIZE // 2, MAP_SIZE // 2)])
     apple: Coord = field(default_factory=lambda: Coord(0, 0))
     map_size: Tuple[int, int] = (MAP_SIZE, MAP_SIZE)
-    stage: Phase = Phase.START
+    stage: Event = Event.START_MENU
+
